refactor: Improve abstrict directory structure
