+++ .gitignore
@@ -33,6 +33,7 @@ local.properties
 # node.js
 #
 node_modules/
+.vscode
 npm-debug.log
 yarn-error.log
 
@@ -44,11 +45,6 @@ buck-out/
 
 # fastlane
 #
-# It is recommended to not store the screenshots in the git repo. Instead, use fastlane to re-generate the
-# screenshots whenever they are needed.
-# For more information about the recommended setup visit:
-# https://docs.fastlane.tools/best-practices/source-control/
-
 */fastlane/report.xml
 */fastlane/Preview.html
 */fastlane/screenshots
