+++ src/components/Logo/index.js
@@ -1,4 +1,5 @@
 import React from "react";
+
 import { TouchableOpacity, Text, View, Image, StyleSheet } from "react-native";
 
 /*
