+++ src/components/Modals/index.js
@@ -1,0 +1,4 @@
+import ContentModal from "./ContentModal";
+import InputModalContainer from "./InputModalContainer";
+
+export { InputModalContainer, ContentModal };
