+++ src/components/index.js
@@ -1,15 +1,7 @@
 import DefaultButton from "./DefaultButton";
-import InputModal from "./InputModal/InputModal";
-import InputModalContainer from "./InputModal/InputModalContainer";
 import Logo from "./Logo";
+import InputModalContainer from "./Modals/InputModalContainer";
 import SmallButton from "./SamllButton";
 import SquareButton from "./SquareButton";
 
-export {
-  DefaultButton,
-  SquareButton,
-  SmallButton,
-  Logo,
-  InputModalContainer,
-  InputModal,
-};
+export { DefaultButton, SquareButton, SmallButton, Logo, InputModalContainer };
